Started a Pokemon List Clone
Due to the nature of the data structure format I used, any changes to health of a Pokemon will make those changes to any of that species at current. I need to make it so that this will only affect members on the user's teams. I will worry about preventing duplication later.